+++ lib/opte-test-utils/src/dhcp.rs
@@ -11,10 +11,10 @@ use dhcpv6::protocol::MessageType;
 use opte::engine::dhcp::DHCP_CLIENT_PORT;
 use opte::engine::dhcp::DHCP_SERVER_PORT;
 use opte::engine::dhcpv6;
-use opte::engine::ingot_base::Ethernet;
-use opte::engine::ingot_base::Ipv4;
-use opte::engine::ingot_base::Ipv6;
+use opte::engine::ether::Ethernet;
 use opte::engine::ingot_packet::MsgBlk;
+use opte::engine::ip::v4::Ipv4;
+use opte::engine::ip::v6::Ipv6;
 use opte::ingot::ethernet::Ethertype;
 use opte::ingot::ip::IpProtocol;
 use opte::ingot::udp::Udp;
