+++ lib/opte-test-utils/src/icmp.rs
@@ -7,11 +7,11 @@
 //! Routines for ICMP testing.
 
 use opte::api::*;
-use opte::engine::ingot_base::Ethernet;
-use opte::engine::ingot_base::Ipv4;
-use opte::engine::ingot_base::Ipv6;
-use opte::engine::ingot_base::L3;
+use opte::engine::ether::Ethernet;
 use opte::engine::ingot_packet::MsgBlk;
+use opte::engine::ip::v4::Ipv4;
+use opte::engine::ip::v6::Ipv6;
+use opte::engine::ip::L3;
 use opte::ingot::ethernet::Ethertype;
 use opte::ingot::ip::IpProtocol as IngotIpProto;
 use opte::ingot::types::HeaderLen;
