+++ lib/opte-test-utils/src/lib.rs
@@ -20,26 +20,26 @@ pub use opte::api::Direction::*;
 pub use opte::api::MacAddr;
 pub use opte::engine::ether::EtherMeta;
 pub use opte::engine::ether::EtherType;
+pub use opte::engine::ether::Ethernet;
 pub use opte::engine::geneve::GeneveMeta;
 pub use opte::engine::geneve::GeneveOption;
 pub use opte::engine::geneve::OxideOption;
 pub use opte::engine::geneve::Vni;
-use opte::engine::geneve::GENEVE_OPT_CLASS_OXIDE;
-use opte::engine::geneve::GENEVE_PORT;
+pub use opte::engine::geneve::GENEVE_OPT_CLASS_OXIDE;
+pub use opte::engine::geneve::GENEVE_PORT;
 pub use opte::engine::headers::IpAddr;
 pub use opte::engine::headers::IpCidr;
-use opte::engine::ingot_base::Ethernet;
-use opte::engine::ingot_base::Ipv4;
-use opte::engine::ingot_base::Ipv6;
-use opte::engine::ingot_base::L3Repr;
-use opte::engine::ingot_packet::MblkLiteParsed;
-use opte::engine::ingot_packet::MsgBlk;
-use opte::engine::ingot_packet::Packet2;
+pub use opte::engine::ingot_packet::MblkLiteParsed;
+pub use opte::engine::ingot_packet::MsgBlk;
+pub use opte::engine::ingot_packet::Packet2;
+pub use opte::engine::ip::v4::Ipv4;
+pub use opte::engine::ip::v6::Ipv6;
+pub use opte::engine::ip::L3Repr;
 pub use opte::engine::ip4::Ipv4Addr;
 pub use opte::engine::ip4::Protocol;
 pub use opte::engine::ip6::Ipv6Addr;
 pub use opte::engine::layer::DenyReason;
-use opte::engine::packet::ParseError;
+pub use opte::engine::packet::ParseError;
 pub use opte::engine::port::meta::ActionMeta;
 pub use opte::engine::port::DropReason;
 pub use opte::engine::port::Port;
@@ -47,18 +47,18 @@ pub use opte::engine::port::PortBuilder;
 pub use opte::engine::port::ProcessResult;
 pub use opte::engine::port::ProcessResult::*;
 pub use opte::engine::GenericUlp;
-use opte::engine::NetworkParser;
+pub use opte::engine::NetworkParser;
 pub use opte::ingot::ethernet::Ethertype;
-use opte::ingot::geneve::Geneve;
-use opte::ingot::geneve::GeneveOpt;
-use opte::ingot::geneve::GeneveOptionType;
+pub use opte::ingot::geneve::Geneve;
+pub use opte::ingot::geneve::GeneveOpt;
+pub use opte::ingot::geneve::GeneveOptionType;
 pub use opte::ingot::ip::IpProtocol as IngotIpProto;
-use opte::ingot::tcp::Tcp;
-use opte::ingot::tcp::TcpFlags as IngotTcpFlags;
-use opte::ingot::types::Emit;
-use opte::ingot::types::EmitDoesNotRelyOnBufContents;
-use opte::ingot::types::HeaderLen;
-use opte::ingot::udp::Udp;
+pub use opte::ingot::tcp::Tcp;
+pub use opte::ingot::tcp::TcpFlags as IngotTcpFlags;
+pub use opte::ingot::types::Emit;
+pub use opte::ingot::types::EmitDoesNotRelyOnBufContents;
+pub use opte::ingot::types::HeaderLen;
+pub use opte::ingot::udp::Udp;
 pub use opte::ExecCtx;
 pub use oxide_vpc::api::AddFwRuleReq;
 pub use oxide_vpc::api::DhcpCfg;
@@ -82,8 +82,8 @@ pub use oxide_vpc::engine::overlay;
 pub use oxide_vpc::engine::overlay::Virt2Boundary;
 pub use oxide_vpc::engine::overlay::Virt2Phys;
 pub use oxide_vpc::engine::overlay::VpcMappings;
-use oxide_vpc::engine::overlay::BOUNDARY_SERVICES_VNI;
-use oxide_vpc::engine::overlay::TUNNEL_ENDPOINT_MAC;
+pub use oxide_vpc::engine::overlay::BOUNDARY_SERVICES_VNI;
+pub use oxide_vpc::engine::overlay::TUNNEL_ENDPOINT_MAC;
 pub use oxide_vpc::engine::router;
 pub use oxide_vpc::engine::VpcNetwork;
 pub use oxide_vpc::engine::VpcParser;
