+++ lib/opte/src/engine/dhcpv6/protocol.rs
@@ -21,11 +21,11 @@ use crate::engine::dhcpv6::ALL_RELAYS_AND_SERVERS;
 use crate::engine::dhcpv6::ALL_SERVERS;
 use crate::engine::dhcpv6::CLIENT_PORT;
 use crate::engine::dhcpv6::SERVER_PORT;
-use crate::engine::ingot_base::Ethernet;
-use crate::engine::ingot_base::Ipv6;
-use crate::engine::ingot_base::Ipv6Ref;
+use crate::engine::ether::Ethernet;
 use crate::engine::ingot_packet::MblkPacketData;
 use crate::engine::ingot_packet::MsgBlk;
+use crate::engine::ip::v6::Ipv6;
+use crate::engine::ip::v6::Ipv6Ref;
 use crate::engine::predicate::DataPredicate;
 use crate::engine::predicate::EtherAddrMatch;
 use crate::engine::predicate::IpProtoMatch;
