+++ lib/opte/src/engine/ingot_packet.rs
@@ -4,11 +4,13 @@
 
 // Copyright 2024 Oxide Computer Company
 
-use super::checksum::Checksum as OpteCsum;
 use super::checksum::Checksum;
-use super::checksum::HeaderChecksum;
 use super::ether::EtherMeta;
 use super::ether::EtherMod;
+use super::ether::Ethernet;
+use super::ether::EthernetMut;
+use super::ether::EthernetPacket;
+use super::ether::ValidEthernet;
 use super::geneve::OxideOption;
 use super::geneve::GENEVE_OPT_CLASS_OXIDE;
 use super::geneve::GENEVE_PORT;
@@ -26,27 +28,17 @@ use super::icmp::IcmpEchoMut;
 use super::icmp::IcmpEchoRef;
 use super::icmp::QueryEcho;
 use super::icmp::ValidIcmpEcho;
-use super::ingot_base::Ethernet;
-use super::ingot_base::EthernetMut;
-use super::ingot_base::EthernetPacket;
-use super::ingot_base::EthernetRef;
-use super::ingot_base::Ipv4;
-use super::ingot_base::Ipv4Mut;
-use super::ingot_base::Ipv4Packet;
-use super::ingot_base::Ipv4Ref;
-use super::ingot_base::Ipv6;
-use super::ingot_base::Ipv6Mut;
-use super::ingot_base::Ipv6Packet;
-use super::ingot_base::Ipv6Ref;
-use super::ingot_base::L3Repr;
-use super::ingot_base::Ulp;
-use super::ingot_base::UlpRepr;
-use super::ingot_base::ValidEthernet;
-use super::ingot_base::ValidL3;
-use super::ingot_base::ValidL4;
-use super::ingot_base::ValidUlp;
-use super::ingot_base::L3;
-use super::ingot_base::L4;
+use super::ip::v4::Ipv4;
+use super::ip::v4::Ipv4Mut;
+use super::ip::v4::Ipv4Packet;
+use super::ip::v4::Ipv4Ref;
+use super::ip::v6::Ipv6;
+use super::ip::v6::Ipv6Mut;
+use super::ip::v6::Ipv6Packet;
+use super::ip::v6::Ipv6Ref;
+use super::ip::L3Repr;
+use super::ip::ValidL3;
+use super::ip::L3;
 use super::packet::allocb;
 use super::packet::AddrPair;
 use super::packet::BodyTransform;
@@ -58,6 +50,9 @@ use super::packet::SegAdjustError;
 use super::packet::WrapError;
 use super::packet::WriteError;
 use super::packet::FLOW_ID_DEFAULT;
+use super::parse::NoEncap;
+use super::parse::Ulp;
+use super::parse::UlpRepr;
 use super::rule::CompiledEncap;
 use super::rule::CompiledTransform;
 use super::rule::HdrTransform;
@@ -91,7 +86,6 @@ use ingot::geneve::Geneve;
 use ingot::geneve::GeneveMut;
 use ingot::geneve::GeneveOpt;
 use ingot::geneve::GeneveOptionType;
-use ingot::geneve::GenevePacket;
 use ingot::geneve::GeneveRef;
 use ingot::geneve::ValidGeneve;
 use ingot::icmp::IcmpV4Mut;
@@ -115,7 +109,6 @@ use ingot::types::HeaderLen;
 use ingot::types::HeaderParse;
 use ingot::types::InlineHeader;
 use ingot::types::NextLayer;
-use ingot::types::ParseControl;
 use ingot::types::ParseError as IngotParseErr;
 use ingot::types::Parsed as IngotParsed;
 use ingot::types::Read;
@@ -125,406 +118,12 @@ use ingot::udp::UdpMut;
 use ingot::udp::UdpPacket;
 use ingot::udp::UdpRef;
 use ingot::udp::ValidUdp;
-use ingot::Parse;
 use opte_api::Direction;
 use opte_api::Ipv6Addr;
 use opte_api::Vni;
 use zerocopy::ByteSlice;
 use zerocopy::ByteSliceMut;
 use zerocopy::IntoBytes;
-
-#[derive(Parse)]
-pub struct GeneveOverV6<Q: ByteSlice> {
-    pub outer_eth: EthernetPacket<Q>,
-    #[ingot(from = "L3<Q>")]
-    pub outer_v6: Ipv6Packet<Q>,
-    #[ingot(from = "L4<Q>", control = geneve_dst_port)]
-    pub outer_udp: UdpPacket<Q>,
-    pub outer_encap: GenevePacket<Q>,
-
-    pub inner_eth: EthernetPacket<Q>,
-    pub inner_l3: L3<Q>,
-    pub inner_ulp: Ulp<Q>,
-}
-
-#[inline]
-fn geneve_dst_port<V: ByteSlice>(l4: &ValidL4<V>) -> ParseControl {
-    match l4 {
-        ValidL4::Udp(u) if u.destination() == GENEVE_PORT => {
-            ParseControl::Continue
-        }
-        _ => ParseControl::Reject,
-    }
-}
-
-#[inline]
-fn exit_on_arp<V: ByteSlice>(eth: &ValidEthernet<V>) -> ParseControl {
-    if eth.ethertype() == Ethertype::ARP {
-        ParseControl::Accept
-    } else {
-        ParseControl::Continue
-    }
-}
-
-#[derive(Parse)]
-pub struct NoEncap<Q: ByteSlice> {
-    #[ingot(control = exit_on_arp)]
-    pub inner_eth: EthernetPacket<Q>,
-    pub inner_l3: Option<L3<Q>>,
-    pub inner_ulp: Option<Ulp<Q>>,
-}
-
-impl<T: ByteSlice> From<ValidNoEncap<T>> for OpteMeta<T> {
-    #[inline]
-    fn from(value: ValidNoEncap<T>) -> Self {
-        NoEncap::from(value).into()
-    }
-}
-
-impl<V: ByteSliceMut> LightweightMeta<V> for ValidNoEncap<V> {
-    #[inline]
-    fn flow(&self) -> InnerFlowId {
-        let (proto, addrs) = match &self.inner_l3 {
-            Some(ValidL3::Ipv4(pkt)) => (
-                pkt.protocol().0,
-                AddrPair::V4 { src: pkt.source(), dst: pkt.destination() },
-            ),
-            Some(ValidL3::Ipv6(pkt)) => (
-                pkt.next_layer().unwrap_or_default().0,
-                AddrPair::V6 { src: pkt.source(), dst: pkt.destination() },
-            ),
-            None => (255, FLOW_ID_DEFAULT.addrs),
-        };
-
-        let (src_port, dst_port) = self
-            .inner_ulp
-            .as_ref()
-            .map(|ulp| {
-                (
-                    actual_src_port_v(ulp)
-                        .or_else(|| pseudo_port_v(ulp))
-                        .unwrap_or(0),
-                    actual_dst_port_v(ulp)
-                        .or_else(|| pseudo_port_v(ulp))
-                        .unwrap_or(0),
-                )
-            })
-            .unwrap_or((0, 0));
-
-        InnerFlowId { proto: proto.into(), addrs, src_port, dst_port }
-    }
-
-    #[inline]
-    fn run_compiled_transform(&mut self, transform: &CompiledTransform)
-    where
-        V: ByteSliceMut,
-    {
-        // TODO: break out commonalities for this and geneve.
-        if let Some(ether_tx) = &transform.inner_ether {
-            if let Some(new_src) = &ether_tx.src {
-                self.inner_eth.set_source(*new_src);
-            }
-            if let Some(new_dst) = &ether_tx.dst {
-                self.inner_eth.set_destination(*new_dst);
-            }
-        }
-        match (&mut self.inner_l3, &transform.inner_ip) {
-            (Some(ValidL3::Ipv4(pkt)), Some(IpMod::Ip4(tx))) => {
-                if let Some(new_src) = &tx.src {
-                    pkt.set_source(*new_src);
-                }
-                if let Some(new_dst) = &tx.dst {
-                    pkt.set_destination(*new_dst);
-                }
-                if let Some(new_proto) = &tx.proto {
-                    pkt.set_protocol(IpProtocol(u8::from(*new_proto)));
-                }
-            }
-            (Some(ValidL3::Ipv6(pkt)), Some(IpMod::Ip6(tx))) => {
-                if let Some(new_src) = &tx.src {
-                    pkt.set_source(*new_src);
-                }
-                if let Some(new_dst) = &tx.dst {
-                    pkt.set_destination(*new_dst);
-                }
-                if let Some(new_proto) = &tx.proto {
-                    // TODO: wrong in the face of EHs...
-                    // For now, we never use this on our dataplane.
-                    pkt.set_next_header(IpProtocol(u8::from(*new_proto)));
-                }
-            }
-            _ => {}
-        }
-
-        match (&mut self.inner_ulp, &transform.inner_ulp) {
-            (Some(ValidUlp::Tcp(pkt)), Some(tx)) => {
-                if let Some(flags) = tx.tcp_flags {
-                    pkt.set_flags(TcpFlags::from_bits_retain(flags));
-                }
-
-                if let Some(new_src) = &tx.generic.src_port {
-                    pkt.set_source(*new_src);
-                }
-
-                if let Some(new_dst) = &tx.generic.dst_port {
-                    pkt.set_destination(*new_dst);
-                }
-            }
-            (Some(ValidUlp::Udp(pkt)), Some(tx)) => {
-                if let Some(new_src) = &tx.generic.src_port {
-                    pkt.set_source(*new_src);
-                }
-
-                if let Some(new_dst) = &tx.generic.dst_port {
-                    pkt.set_destination(*new_dst);
-                }
-            }
-            (Some(ValidUlp::IcmpV4(pkt)), Some(tx))
-                if pkt.ty() == 0 || pkt.ty() == 8 =>
-            {
-                if let Some(new_id) = tx.icmp_id {
-                    pkt.rest_of_hdr_mut()[..2]
-                        .copy_from_slice(&new_id.to_be_bytes())
-                }
-            }
-            (Some(ValidUlp::IcmpV6(pkt)), Some(tx))
-                if pkt.ty() == 128 || pkt.ty() == 129 =>
-            {
-                if let Some(new_id) = tx.icmp_id {
-                    pkt.rest_of_hdr_mut()[..2]
-                        .copy_from_slice(&new_id.to_be_bytes())
-                }
-            }
-            _ => {}
-        }
-    }
-
-    // FIXME: identical to Geneve.
-    #[inline]
-    fn compute_body_csum(&self) -> Option<OpteCsum> {
-        let use_pseudo = if let Some(v) = &self.inner_ulp {
-            !matches!(v, ValidUlp::IcmpV4(_))
-        } else {
-            false
-        };
-
-        let pseudo_csum = match self.inner_eth.ethertype() {
-            Ethertype::IPV4 | Ethertype::IPV6 => {
-                self.inner_l3.as_ref().map(|v| v.pseudo_header())
-            }
-            // Includes ARP.
-            _ => return None,
-        };
-
-        let Some(pseudo_csum) = pseudo_csum else {
-            return None;
-        };
-
-        self.inner_ulp.as_ref().and_then(csum_minus_hdr).map(|mut v| {
-            if use_pseudo {
-                v -= pseudo_csum;
-            }
-            v
-        })
-    }
-
-    #[inline]
-    fn encap_len(&self) -> u16 {
-        0
-    }
-
-    #[inline]
-    fn update_inner_checksums(&mut self, body_csum: OpteCsum) {
-        if let Some(l3) = self.inner_l3.as_mut() {
-            if let Some(ulp) = self.inner_ulp.as_mut() {
-                ulp.compute_checksum(body_csum, l3);
-            }
-            l3.compute_checksum();
-        }
-    }
-
-    #[inline]
-    fn inner_tcp(&self) -> Option<&impl TcpRef<V>> {
-        match self.inner_ulp.as_ref() {
-            Some(ValidUlp::Tcp(t)) => Some(t),
-            _ => None,
-        }
-    }
-}
-
-impl<T: ByteSlice> From<ValidGeneveOverV6<T>> for OpteMeta<T> {
-    #[inline]
-    fn from(value: ValidGeneveOverV6<T>) -> Self {
-        OpteMeta {
-            outer_eth: Some(value.outer_eth.into()),
-            outer_l3: Some(L3::Ipv6(value.outer_v6.into())),
-            outer_encap: Some(InlineHeader::Raw(ValidEncapMeta::Geneve(
-                value.outer_udp,
-                value.outer_encap,
-            ))),
-            inner_eth: value.inner_eth.into(),
-            inner_l3: Some(value.inner_l3.into()),
-            inner_ulp: Some(value.inner_ulp.into()),
-        }
-    }
-}
-
-impl<V: ByteSliceMut> LightweightMeta<V> for ValidGeneveOverV6<V> {
-    #[inline]
-    fn flow(&self) -> InnerFlowId {
-        let (proto, addrs) = match &self.inner_l3 {
-            ValidL3::Ipv4(pkt) => (
-                pkt.protocol().0,
-                AddrPair::V4 { src: pkt.source(), dst: pkt.destination() },
-            ),
-            ValidL3::Ipv6(pkt) => (
-                pkt.next_layer().unwrap_or_default().0,
-                AddrPair::V6 { src: pkt.source(), dst: pkt.destination() },
-            ),
-        };
-
-        let src_port = actual_src_port_v(&self.inner_ulp)
-            .or_else(|| pseudo_port_v(&self.inner_ulp))
-            .unwrap_or(0);
-
-        let dst_port = actual_dst_port_v(&self.inner_ulp)
-            .or_else(|| pseudo_port_v(&self.inner_ulp))
-            .unwrap_or(0);
-
-        InnerFlowId { proto: proto.into(), addrs, src_port, dst_port }
-    }
-
-    #[inline]
-    fn run_compiled_transform(&mut self, transform: &CompiledTransform)
-    where
-        V: ByteSliceMut,
-    {
-        // TODO: break out commonalities for this and geneve.
-        if let Some(ether_tx) = &transform.inner_ether {
-            if let Some(new_src) = &ether_tx.src {
-                self.inner_eth.set_source(*new_src);
-            }
-            if let Some(new_dst) = &ether_tx.dst {
-                self.inner_eth.set_destination(*new_dst);
-            }
-        }
-        match (&mut self.inner_l3, &transform.inner_ip) {
-            (ValidL3::Ipv4(pkt), Some(IpMod::Ip4(tx))) => {
-                if let Some(new_src) = &tx.src {
-                    pkt.set_source(*new_src);
-                }
-                if let Some(new_dst) = &tx.dst {
-                    pkt.set_destination(*new_dst);
-                }
-                if let Some(new_proto) = &tx.proto {
-                    pkt.set_protocol(IpProtocol(u8::from(*new_proto)));
-                }
-            }
-            (ValidL3::Ipv6(pkt), Some(IpMod::Ip6(tx))) => {
-                if let Some(new_src) = &tx.src {
-                    pkt.set_source(*new_src);
-                }
-                if let Some(new_dst) = &tx.dst {
-                    pkt.set_destination(*new_dst);
-                }
-                if let Some(new_proto) = &tx.proto {
-                    // TODO: wrong in the face of EHs...
-                    // For now, we never use this on our dataplane.
-                    pkt.set_next_header(IpProtocol(u8::from(*new_proto)));
-                }
-            }
-            _ => {}
-        }
-
-        match (&mut self.inner_ulp, &transform.inner_ulp) {
-            (ValidUlp::Tcp(pkt), Some(tx)) => {
-                if let Some(flags) = tx.tcp_flags {
-                    pkt.set_flags(TcpFlags::from_bits_retain(flags));
-                }
-
-                if let Some(new_src) = &tx.generic.src_port {
-                    pkt.set_source(*new_src);
-                }
-
-                if let Some(new_dst) = &tx.generic.dst_port {
-                    pkt.set_destination(*new_dst);
-                }
-            }
-            (ValidUlp::Udp(pkt), Some(tx)) => {
-                if let Some(new_src) = &tx.generic.src_port {
-                    pkt.set_source(*new_src);
-                }
-
-                if let Some(new_dst) = &tx.generic.dst_port {
-                    pkt.set_destination(*new_dst);
-                }
-            }
-            (ValidUlp::IcmpV4(pkt), Some(tx))
-                if pkt.ty() == 0 || pkt.ty() == 8 =>
-            {
-                if let Some(new_id) = tx.icmp_id {
-                    pkt.rest_of_hdr_mut()[..2]
-                        .copy_from_slice(&new_id.to_be_bytes())
-                }
-            }
-            (ValidUlp::IcmpV6(pkt), Some(tx))
-                if pkt.ty() == 128 || pkt.ty() == 129 =>
-            {
-                if let Some(new_id) = tx.icmp_id {
-                    pkt.rest_of_hdr_mut()[..2]
-                        .copy_from_slice(&new_id.to_be_bytes())
-                }
-            }
-            _ => {}
-        }
-    }
-
-    #[inline]
-    fn compute_body_csum(&self) -> Option<OpteCsum> {
-        let use_pseudo = !matches!(self.inner_ulp, ValidUlp::IcmpV4(_));
-
-        let pseudo_csum = match self.inner_eth.ethertype() {
-            Ethertype::IPV4 | Ethertype::IPV6 => {
-                Some(self.inner_l3.pseudo_header())
-            }
-            // Includes ARP.
-            _ => return None,
-        };
-
-        let Some(pseudo_csum) = pseudo_csum else {
-            return None;
-        };
-
-        csum_minus_hdr(&self.inner_ulp).map(|mut v| {
-            if use_pseudo {
-                v -= pseudo_csum;
-            }
-            v
-        })
-    }
-
-    #[inline]
-    fn encap_len(&self) -> u16 {
-        (self.outer_eth.packet_length()
-            + self.outer_v6.packet_length()
-            + self.outer_udp.packet_length()
-            + self.outer_encap.packet_length()) as u16
-    }
-
-    #[inline]
-    fn update_inner_checksums(&mut self, body_csum: OpteCsum) {
-        self.inner_ulp.compute_checksum(body_csum, &self.inner_l3);
-        self.inner_l3.compute_checksum();
-    }
-
-    #[inline]
-    fn inner_tcp(&self) -> Option<&impl TcpRef<V>> {
-        match &self.inner_ulp {
-            ValidUlp::Tcp(t) => Some(t),
-            _ => None,
-        }
-    }
-}
 
 /// An individual illumos `mblk_t` -- a single bytestream
 /// comprised of a linked list of data segments.
@@ -1565,70 +1164,6 @@ impl<T: Read> PacketData<T> {
     }
 }
 
-fn actual_src_port<T: ByteSlice>(chunk: &Ulp<T>) -> Option<u16> {
-    match chunk {
-        Ulp::Tcp(pkt) => Some(pkt.source()),
-        Ulp::Udp(pkt) => Some(pkt.source()),
-        _ => None,
-    }
-}
-
-fn actual_src_port_v<T: ByteSlice>(chunk: &ValidUlp<T>) -> Option<u16> {
-    match chunk {
-        ValidUlp::Tcp(pkt) => Some(pkt.source()),
-        ValidUlp::Udp(pkt) => Some(pkt.source()),
-        _ => None,
-    }
-}
-
-fn actual_dst_port<T: ByteSlice>(chunk: &Ulp<T>) -> Option<u16> {
-    match chunk {
-        Ulp::Tcp(pkt) => Some(pkt.destination()),
-        Ulp::Udp(pkt) => Some(pkt.destination()),
-        _ => None,
-    }
-}
-
-fn actual_dst_port_v<T: ByteSlice>(chunk: &ValidUlp<T>) -> Option<u16> {
-    match chunk {
-        ValidUlp::Tcp(pkt) => Some(pkt.destination()),
-        ValidUlp::Udp(pkt) => Some(pkt.destination()),
-        _ => None,
-    }
-}
-
-fn pseudo_port<T: ByteSlice>(chunk: &Ulp<T>) -> Option<u16> {
-    match chunk {
-        Ulp::IcmpV4(pkt)
-            if pkt.code() == 0 && (pkt.ty() == 0 || pkt.ty() == 8) =>
-        {
-            Some(u16::from_be_bytes(pkt.rest_of_hdr()[..2].try_into().unwrap()))
-        }
-        Ulp::IcmpV6(pkt)
-            if pkt.code() == 0 && (pkt.ty() == 128 || pkt.ty() == 129) =>
-        {
-            Some(u16::from_be_bytes(pkt.rest_of_hdr()[..2].try_into().unwrap()))
-        }
-        _ => None,
-    }
-}
-
-fn pseudo_port_v<T: ByteSlice>(chunk: &ValidUlp<T>) -> Option<u16> {
-    match chunk {
-        ValidUlp::IcmpV4(pkt)
-            if pkt.code() == 0 && (pkt.ty() == 0 || pkt.ty() == 8) =>
-        {
-            Some(u16::from_be_bytes(pkt.rest_of_hdr()[..2].try_into().unwrap()))
-        }
-        ValidUlp::IcmpV6(pkt)
-            if pkt.code() == 0 && (pkt.ty() == 128 || pkt.ty() == 129) =>
-        {
-            Some(u16::from_be_bytes(pkt.rest_of_hdr()[..2].try_into().unwrap()))
-        }
-        _ => None,
-    }
-}
-
 impl<T: Read> From<&PacketData<T>> for InnerFlowId {
     #[inline]
     fn from(meta: &PacketData<T>) -> Self {
@@ -1648,11 +1183,11 @@ impl<T: Read> From<&PacketData<T>> for InnerFlowId {
             .inner_ulp()
             .map(|ulp| {
                 (
-                    actual_src_port(ulp)
-                        .or_else(|| pseudo_port(ulp))
+                    ulp.true_src_port()
+                        .or_else(|| ulp.pseudo_port())
                         .unwrap_or(0),
-                    actual_dst_port(ulp)
-                        .or_else(|| pseudo_port(ulp))
+                    ulp.true_dst_port()
+                        .or_else(|| ulp.pseudo_port())
                         .unwrap_or(0),
                 )
             })
@@ -2415,75 +1950,6 @@ impl<T: Read, M: LightweightMeta<T::Chunk>> LiteParsed<T, M> {}
 pub type MblkPacketData<'a> = PacketData<MsgBlkIterMut<'a>>;
 pub type MblkFullParsed<'a> = FullParsed<MsgBlkIterMut<'a>>;
 pub type MblkLiteParsed<'a, M> = LiteParsed<MsgBlkIterMut<'a>, M>;
-
-#[inline]
-fn csum_minus_hdr<V: ByteSlice>(ulp: &ValidUlp<V>) -> Option<Checksum> {
-    match ulp {
-        ValidUlp::IcmpV4(icmp) => {
-            if icmp.checksum() == 0 {
-                return None;
-            }
-
-            let mut csum = OpteCsum::from(HeaderChecksum::wrap(
-                icmp.checksum().to_be_bytes(),
-            ));
-
-            csum.sub_bytes(&[icmp.ty(), icmp.code()]);
-            csum.sub_bytes(icmp.rest_of_hdr_ref());
-
-            Some(csum)
-        }
-        ValidUlp::IcmpV6(icmp) => {
-            if icmp.checksum() == 0 {
-                return None;
-            }
-
-            let mut csum = OpteCsum::from(HeaderChecksum::wrap(
-                icmp.checksum().to_be_bytes(),
-            ));
-
-            csum.sub_bytes(&[icmp.ty(), icmp.code()]);
-            csum.sub_bytes(icmp.rest_of_hdr_ref());
-
-            Some(csum)
-        }
-        ValidUlp::Tcp(tcp) => {
-            if tcp.checksum() == 0 {
-                return None;
-            }
-
-            let mut csum = OpteCsum::from(HeaderChecksum::wrap(
-                tcp.checksum().to_be_bytes(),
-            ));
-
-            let b = tcp.0.as_bytes();
-
-            csum.sub_bytes(&b[0..16]);
-            csum.sub_bytes(&b[18..]);
-
-            csum.sub_bytes(match &tcp.1 {
-                ingot::types::Header::Repr(v) => &v[..],
-                ingot::types::Header::Raw(v) => &v[..],
-            });
-
-            Some(csum)
-        }
-        ValidUlp::Udp(udp) => {
-            if udp.checksum() == 0 {
-                return None;
-            }
-
-            let mut csum = OpteCsum::from(HeaderChecksum::wrap(
-                udp.checksum().to_be_bytes(),
-            ));
-
-            let b = udp.0.as_bytes();
-            csum.sub_bytes(&b[0..6]);
-
-            Some(csum)
-        }
-    }
-}
 
 pub trait QueryLen {
     fn len(&self) -> usize;
