+++ lib/opte/src/engine/predicate.rs
@@ -9,15 +9,15 @@
 use super::dhcp::MessageType as DhcpMessageType;
 use super::dhcpv6::MessageType as Dhcpv6MessageType;
 use super::ether::EtherType;
+use super::ether::EthernetRef;
 use super::icmp::v4::MessageType as IcmpMessageType;
 use super::icmp::v6::MessageType as Icmpv6MessageType;
-use super::ingot_base::EthernetRef;
-use super::ingot_base::Ipv4Ref;
-use super::ingot_base::Ipv6Ref;
-use super::ingot_base::L3;
 use super::ingot_packet::ulp_dst_port;
 use super::ingot_packet::ulp_src_port;
 use super::ingot_packet::MblkPacketData;
+use super::ip::v4::Ipv4Ref;
+use super::ip::v6::Ipv6Ref;
+use super::ip::L3;
 use super::ip4::Ipv4Addr;
 use super::ip4::Ipv4Cidr;
 use super::ip4::Protocol;
