+++ lib/oxide-vpc/src/engine/mod.rs
@@ -18,16 +18,16 @@ use opte::engine::arp::ArpEthIpv4Ref;
 use opte::engine::arp::ArpOp;
 use opte::engine::arp::ValidArpEthIpv4;
 use opte::engine::arp::ARP_HTYPE_ETHERNET;
+use opte::engine::ether::EthernetRef;
 use opte::engine::flow_table::FlowTable;
-use opte::engine::ingot_base::EthernetRef;
 use opte::engine::ingot_packet::FullParsed;
 use opte::engine::ingot_packet::OpteParsed2;
 use opte::engine::ingot_packet::Packet2;
-use opte::engine::ingot_packet::ValidGeneveOverV6;
-use opte::engine::ingot_packet::ValidNoEncap;
 use opte::engine::ip4::Ipv4Addr;
 use opte::engine::packet::InnerFlowId;
 use opte::engine::packet::ParseError;
+use opte::engine::parse::ValidGeneveOverV6;
+use opte::engine::parse::ValidNoEncap;
 use opte::engine::port::UftEntry;
 use opte::engine::Direction;
 use opte::engine::HdlPktAction;
