+++ lib/oxide-vpc/tests/integration_tests.rs
@@ -23,21 +23,21 @@ use opte::engine::arp::ArpEthIpv4Ref;
 use opte::engine::arp::ValidArpEthIpv4;
 use opte::engine::arp::ARP_HTYPE_ETHERNET;
 use opte::engine::dhcpv6;
+use opte::engine::ether::Ethernet;
+use opte::engine::ether::EthernetRef;
 use opte::engine::flow_table::FLOW_DEF_EXPIRE_SECS;
 use opte::engine::geneve::Vni;
-use opte::engine::ingot_base::Ethernet;
-use opte::engine::ingot_base::EthernetRef;
-use opte::engine::ingot_base::Ipv4Ref;
-use opte::engine::ingot_base::Ipv6;
-use opte::engine::ingot_base::Ipv6Ref;
-use opte::engine::ingot_base::ValidL3;
-use opte::engine::ingot_base::ValidUlp;
-use opte::engine::ingot_base::L3;
 use opte::engine::ingot_packet::MblkFullParsed;
 use opte::engine::ingot_packet::MsgBlk;
 use opte::engine::ingot_packet::Packet2;
+use opte::engine::ip::v4::Ipv4Ref;
+use opte::engine::ip::v6::Ipv6;
+use opte::engine::ip::v6::Ipv6Ref;
+use opte::engine::ip::ValidL3;
+use opte::engine::ip::L3;
 use opte::engine::ip4::Ipv4Addr;
 use opte::engine::packet::InnerFlowId;
+use opte::engine::parse::ValidUlp;
 use opte::engine::port::ProcessError;
 use opte::engine::tcp::TcpState;
 use opte::engine::tcp::TIME_WAIT_EXPIRE_SECS;
